server/requirements.txt
fastapi
uvicorn[standard]
pydantic
aiohttp
transformers>=4.0.0
torch>=1.7.0; platform_system != 'Windows'
google-cloud-firestore
